Consolidate DDMv2 wire types into a single file
The frozen DDMv2 protocol types (UpdateV2, PullResponseV2,
UnderlayUpdateV2, TunnelUpdateV2, PathVectorV2, TunnelOriginV2,
IpPrefix, Ipv4Prefix, Ipv6Prefix) were scattered across
ddm-api-types, mg-common, and ddm/src/exchange/mod.rs. Collect
them all into a single dedicated module with a module-level
comment making the freeze guarantee impossible to miss.

Add an expectorate snapshot test over the JSON schema for the two
top-level message types (UpdateV2, PullResponseV2) so that any
accidental change to the frozen wire format fails loudly.

Also remove the DDMv1 conversion helpers (UpdateV1 and its From
impls) that are no longer needed.
