+++ Cargo.toml
@@ -87,7 +87,8 @@ thiserror = "2.0"
 dropshot = { version = "0.17.0", features = [ "usdt-probes" ] }
 dropshot-api-manager = "0.7.2"
 dropshot-api-manager-types = "0.7.2"
-schemars = { version = "0.8", features = [ "uuid1", "chrono" ] }
+expectorate = "1.2.0"
+schemars = { version = "0.8.22", features = [ "uuid1", "chrono" ] }
 tokio = { version = "1.49", features = ["full"] }
 serde_repr = "0.1"
 anyhow = "1.0.102"
