+++ ddm-api-types/versions/src/initial/net.rs
@@ -5,7 +5,7 @@
 use oxnet::IpNet;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::net::{Ipv4Addr, Ipv6Addr};
+use std::net::Ipv6Addr;
 
 #[derive(
     Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
@@ -16,28 +16,4 @@ pub struct TunnelOrigin {
     pub vni: u32,
     #[serde(default)]
     pub metric: u64,
-}
-
-#[derive(
-    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
-)]
-pub struct Ipv6Prefix {
-    pub addr: Ipv6Addr,
-    pub len: u8,
-}
-
-#[derive(
-    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
-)]
-pub struct Ipv4Prefix {
-    pub addr: Ipv4Addr,
-    pub len: u8,
-}
-
-#[derive(
-    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
-)]
-pub enum IpPrefix {
-    V4(Ipv4Prefix),
-    V6(Ipv6Prefix),
 }
