+++ ddm/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dev-dependencies]
 pretty_assertions.workspace = true
+expectorate.workspace = true
 
 [dependencies]
 slog.workspace = true
