+++ ddm/src/exchange/ddm_v2.rs
@@ -1,0 +1,105 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! ALL TYPES IN THIS FILE ARE FOR DDM PROTOCOL VERSION 2. THEY SHALL NEVER
+//! CHANGE. THESE TYPES CAN BE REMOVED WHEN DDMV2 CLIENTS AND SERVERS NO LONGER
+//! EXIST BUT THEIR DEFINITIONS SHALL NEVER CHANGE.
+
+use std::{
+    collections::HashSet,
+    net::{Ipv4Addr, Ipv6Addr},
+};
+
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
+pub struct UpdateV2 {
+    pub underlay: Option<UnderlayUpdateV2>,
+    pub tunnel: Option<TunnelUpdateV2>,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
+pub struct TunnelUpdateV2 {
+    pub announce: HashSet<TunnelOriginV2>,
+    pub withdraw: HashSet<TunnelOriginV2>,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
+pub struct PullResponseV2 {
+    pub underlay: Option<HashSet<PathVectorV2>>,
+    pub tunnel: Option<HashSet<TunnelOriginV2>>,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
+pub struct UnderlayUpdateV2 {
+    pub announce: HashSet<PathVectorV2>,
+    pub withdraw: HashSet<PathVectorV2>,
+}
+
+#[derive(
+    Debug, Clone, PartialEq, Eq, Hash, Deserialize, Serialize, JsonSchema,
+)]
+pub struct PathVectorV2 {
+    pub destination: Ipv6Prefix,
+    pub path: Vec<String>,
+}
+
+#[derive(
+    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
+)]
+pub struct Ipv6Prefix {
+    pub addr: Ipv6Addr,
+    pub len: u8,
+}
+
+#[derive(
+    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
+)]
+pub struct Ipv4Prefix {
+    pub addr: Ipv4Addr,
+    pub len: u8,
+}
+
+#[derive(
+    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
+)]
+pub enum IpPrefix {
+    V4(Ipv4Prefix),
+    V6(Ipv6Prefix),
+}
+
+#[derive(
+    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
+)]
+pub struct TunnelOriginV2 {
+    pub overlay_prefix: IpPrefix,
+    pub boundary_addr: Ipv6Addr,
+    pub vni: u32,
+    #[serde(default)]
+    pub metric: u64,
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    // Write out the JSON schema for the DDMv2 protocol to a file for
+    // validation. This should not change.
+    #[test]
+    fn test_ddm_v2_protocol() {
+        #[derive(JsonSchema)]
+        #[allow(dead_code)]
+        struct Protocol {
+            update: UpdateV2,
+            pull_response: PullResponseV2,
+        }
+
+        let schema = schemars::schema_for!(Protocol);
+        expectorate::assert_contents(
+            "tests/output/ddm_v2_protocol.json",
+            &serde_json::to_string_pretty(&schema).unwrap(),
+        );
+    }
+}
